feat: rename hooks and remove default exports
