--- src/hooks/useSubscribeModal.ts
@@ -1,15 +1,0 @@
-import { create } from 'zustand';
-
-interface SubscribeModalStore {
-  isOpen: boolean;
-  onOpen: () => void;
-  onClose: () => void;
-}
-
-const useSubscribeModal = create<SubscribeModalStore>((set) => ({
-  isOpen: false,
-  onOpen: () => set({ isOpen: true }),
-  onClose: () => set({ isOpen: false }),
-}));
-
-export default useSubscribeModal;
